Fail if we cannot make output dirs
Failing to create either of these two dirs should result in an immediate exit, as downstream code assumes that they've been created and will fail ungracefully, with a red-herring error.

Further, there are many reasons why `mkdir` can fail, not just because the "Folder already exists". For example, permissions issues can cause a failure, in which case the print statement is misleading.